exfat: remove unneeded type cast

+++ exfat_fs.h
@@ -500,7 +500,7 @@ extern int exfat_find_dir_entry(struct super_block *sb,
 		struct exfat_uni_name *p_uniname, int num_entries,
 		struct exfat_dos_name *unused, unsigned int type);
 extern int exfat_zeroed_cluster(struct super_block *sb,
-		unsigned long long blknr, unsigned long long num_secs);
+		unsigned long long blknr, unsigned int num_secs);
 extern int exfat_alloc_new_dir(struct inode *inode, struct exfat_chain *clu);
 
 /* inode.c */
